Fix #29: Lose Criteria, for custom query build; context match uses regex
The Criteria class for Solr queries in Spring is not really fit for
purpose, in several regards. Most importantly, it does not accommodate
regular expressions, even though the Lucene engine can handle them. This
is expressly problematic for the requisite exact context matching, as a
regex is required to exclude entries containing terms which are absent
from the query. Apart from that, Criteria appears to implement the
toString() method as an afterthought at best: whitespace is missing
around joining operations and, more brokenly still, a joining "AND" is
prepended to the return value, making the Criteria class largely useless
even as a query sub-string generator (e.g. for incorporation into a
larger query which does contain one or more regular expressions). Better
to write a custom query builder, as per this commit.